#684: refactor 提取公共分块/解析函数到shared.js，三个handler复用
